requirements.txt
asgiref==3.5.0
Django==3.2.12
django-ckeditor==6.2.0
django-js-asset==2.0.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.1.1